% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploding_boxplot.R
\name{exploding_boxplot}
\alias{exploding_boxplot}
\title{<Add Title...>}
\usage{
exploding_boxplot(data, y = NULL, group = NULL, color = NULL,
  label = NULL, iqr = NULL, margin = NULL, xlab = NULL, ylab = NULL,
  width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}

